Makes buffers retain the backing device allocator.
Since the device allocator often lives for the life of whatever is being done, this was not being caught. But some Python uses were hitting it. Added a Python test to catch regressions at that level (which should also catch various Python ref-counting issues). I'm not sure if there is a great place to put a C test of this kind.

Fixes #9033